+++ src/components/landing/NoisyRectangle.tsx
