+++ src/components/proofs/Line.tsx
