+++ src/components/proofs/ReadyProofsList.tsx
@@ -1,7 +1,7 @@
 import { BodyText } from 'components/Text'
 import { Suspense } from 'react'
 import { useSnapshot } from 'valtio'
-import ContractListContainer from 'components/ContractListContainer'
+import ContractListContainer from 'components/proofs/ContractListContainer'
 import Proof from 'components/proofs/Proof'
 import ProofStore from 'stores/ProofStore'
